# glowzi-backend/glowzi-identity-service/src/main/java/com/glowzi/identity/interfaces/rest/dto/RegisterUserRequest.java
package com.glowzi.identity.interfaces.rest.dto;

import com.glowzi.identity.domain.enums.UserRole;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Pattern;
import jakarta.validation.constraints.Size;

public class RegisterUserRequest {

    @NotBlank
    @Size(max = 100, message = "Full name must be at most 100 characters")
    private String fullName;

    @NotBlank
    @Pattern(regexp = "^\\+[1-9]\\d{6,14}$",
             message = "Phone must be in E.164 format (e.g. +966501234567)")
    private String phone;

    @NotBlank
    @Size(min = 8, max = 64, message = "Password must be between 8 and 64 characters")
    @Pattern(regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).+$",
             message = "Password must contain at least one uppercase letter, one lowercase letter, and one digit")
    private String password;

    @NotNull
    private UserRole role;

    private String preferredLanguage;

    public String getFullName() { return fullName; }
    public void setFullName(String fullName) { this.fullName = fullName; }

    public String getPhone() { return phone; }
    public void setPhone(String phone) { this.phone = phone; }

    public String getPassword() { return password; }
    public void setPassword(String password) { this.password = password; }

    public UserRole getRole() { return role; }
    public void setRole(UserRole role) { this.role = role; }

    public String getPreferredLanguage() { return preferredLanguage; }
    public void setPreferredLanguage(String preferredLanguage) { this.preferredLanguage = preferredLanguage; }
}