# glowzi-backend/glowzi-identity-service/src/main/java/com/glowzi/identity/interfaces/rest/dto/TokenResponse.java
package com.glowzi.identity.interfaces.rest.dto;

/**
 * Response body for POST /auth/refresh.
 * Returns a new access token + new refresh token.
 */
public class TokenResponse {

    private final String accessToken;
    private final String refreshToken;
    private final long expiresIn;   // seconds until the access token expires

    public TokenResponse(String accessToken, String refreshToken, long expiresIn) {
        this.accessToken = accessToken;
        this.refreshToken = refreshToken;
        this.expiresIn = expiresIn;
    }

    public String getAccessToken()  { return accessToken; }
    public String getRefreshToken() { return refreshToken; }
    public long getExpiresIn()      { return expiresIn; }
}